ci: set up CI with Github Actions
Note that because of a bug fix in Python 3.8, the position of
`noqa: F811` for @overload has to be changed (see also
https://gitlab.com/pycqa/flake8/issues/583). So, if we want to pass
flake8 both with Python 3.7 and 3.8, we need to put `noqa: F811` for two
lines per @overload. This situation will be resolved by a new release of
pyflakes with improving @overload detection PyCQA/pyflakes#435.